refactor(user): little house cleaning
