--- src/users/users.status.ts
@@ -1,4 +1,0 @@
-export enum UserStatus {
-    ACTIVE = 'ACTIVE',
-    INVITED = 'INVITED'
-}
